# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Grantex Auth Service API
  version: 0.5.0
  description: |
    Grantex is an open-source delegated authorization protocol and reference implementation for AI agents. It gives each agent a verifiable identity and scoped, time-limited, revocable authority from a human or organization, with multi-agent delegation, service-side verification, and audit records.

    Grantex is owned by Orchestrum Technologies LLP. Inventor and owner: Sanjeev Kumar. Ownership contact: sanjeev@orchestrum.in or mishra.sanjeev@gmail.com.

    The Grantex Auth Service implements this protocol for API clients.
    It provides agent registration, authorization flows, token management, audit logging, policy enforcement,
    anomaly detection, compliance exports, SCIM provisioning, SSO, and billing.

    ## Authentication
    Most endpoints require a **Bearer token** (your developer API key) in the `Authorization` header.
    SCIM 2.0 user endpoints use a separate SCIM Bearer token.
    Public endpoints (health, JWKS, consent, SSO flow) require no authentication.

    ## Rate Limits
    - **Default Fastify IP policy (routes without an override)**: 5,000 requests/minute per source IP
    - **Standard developer API-key - Free plan**: 100 requests/minute per developer
    - **Standard developer API-key - Pro plan**: 500 requests/minute per developer
    - **Standard developer API-key - Enterprise plan**: 2,000 requests/minute per developer
...

...
  license:
    name: Apache 2.0
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Developer API key
    grantTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Grantex grant token issued to an authorized agent
    metricsBearerAuth:
      type: http
      scheme: bearer
      description: METRICS_API_KEY bearer token (required by default in production)
    principalSessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived principal-session token
    dpopAuth:
      type: http
      scheme: DPoP
      bearerFormat: JWT
      description: OAuth Agent Grants access token plus the DPoP proof header
    scimBearer:
      type: http
      scheme: bearer
      description: SCIM provisioning token
    adminAuth:
      type: http
      scheme: bearer
      description: Service administrator credential (ADMIN_API_KEY); not a developer API key
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        statusCode:
          type: integer
      required: [error, code, statusCode]
...
security:
  - bearerAuth: []
paths:
...
  /consent:
    get:
      operationId: getConsentPage
      tags: [Consent]
      summary: Consent UI page
      description: HTML consent page for end-user approval.
      security: []
      parameters:
        - name: req
          in: query
          schema:
            type: string
          description: Authorization request ID
      responses:
        "200":
          description: HTML consent page
          content:
            text/html:
              schema:
                type: string
  /v1/consent/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Authorization request ID
    get:
      operationId: getConsentDetails
      tags: [Consent]
      summary: Get consent request details
      security: []
      responses:
        "200":
          description: Consent request details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  agentName:
                    type: string
                  agentDid:
                    type: string
                  agentDescription:
                    type: string
                  scopes:
                    type: array
                    items:
                      type: string
                  scopeDescriptions:
                    type: array
                    items:
                      type: string
                  expiresAt:
                    type: string
                    format: date-time
                  status:
                    type: string
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Request expired or already processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/consent/{id}/approve:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Authorization request ID
    post:
      operationId: approveConsent
      tags: [Consent]
      summary: Approve consent (end-user action)
      security: []
      responses:
        "200":
          description: Consent approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  redirectUri:
                    type: string
                  state:
                    type: string
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Request expired or already processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/consent/{id}/deny:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Authorization request ID
    post:
      operationId: denyConsent
      tags: [Consent]
      summary: Deny consent (end-user action)
      security: []
      responses:
        "200":
          description: Consent denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirectUri:
                    type: string
                  state:
                    type: string
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Tokens ───────────────────────────────────────────────────────────
...
  /v1/webauthn/assert/options:
    post:
      operationId: webauthnAssertOptions
      tags: [WebAuthn]
      summary: Generate assertion options
      description: |
        Generate FIDO2/WebAuthn assertion options for the consent flow. Returns a challenge and
        PublicKeyCredentialRequestOptions that the browser passes to navigator.credentials.get().
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [principalId, authRequestId]
              properties:
                principalId:
                  type: string
                  description: The end-user's principal ID
                authRequestId:
                  type: string
                  description: The authorization request ID this assertion is for
      responses:
        "200":
          description: Assertion options
          content:
            application/json:
              schema:
                type: object
                properties:
                  challengeId:
                    type: string
                  challenge:
                    type: string
                  allowCredentials:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        type:
                          type: string
                  userVerification:
                    type: string
                    enum: [required]
                  timeout:
                    type: integer
        "400":
          description: Missing parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No credentials registered for this principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/webauthn/assert/verify:
    post:
      operationId: webauthnAssertVerify
      tags: [WebAuthn]
      summary: Verify assertion
      description: |
        Verify the assertion response from navigator.credentials.get(). On success, the associated
        authorization request is approved with FIDO evidence attached to the grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, response]
              properties:
                challengeId:
                  type: string
                response:
                  type: object
                  properties:
                    id:
                      type: string
                    rawId:
                      type: string
                    type:
                      type: string
                    response:
                      type: object
                      properties:
                        clientDataJSON:
                          type: string
                        authenticatorData:
                          type: string
                        signature:
                          type: string
      responses:
        "200":
          description: Assertion verified, grant approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  grantId:
                    type: string
                  fidoEvidence:
                    type: object
                    properties:
                      credentialId:
                        type: string
                      authenticatorType:
                        type: string
                      userVerified:
                        type: boolean
                      assertedAt:
                        type: string
                        format: date-time
        "400":
          description: Invalid assertion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Verifiable Credentials ──────────────────────────────────────────
...